// state/types.ts
export enum GENDER {
  MALE = "Male",
  FEMALE = "Female",
  OTHER = "Other",
}

export enum ACTIVITY_LEVEL {
  MINIMAL = "Little to no exercise",
  LIGHT = "Light exercise a few times a week",
  MODERATE = "Moderate exercise 3-5 times a week",
  HEAVY = "Heavy exercise 6-7 times a week",
}

export enum ACTIVITY_LEVEL_SCORE {
  MINIMAL = 1.2,
  LIGHT = 1.375,
  MODERATE = 1.55,
  HEAVY = 1.725,
}